+++ cmd/shortener/shorty.json
@@ -3,5 +3,6 @@
     "base_url": "http://localhost",
     "file_storage_path": "/tmp/short-url-db.json",
     "database_dsn": "host=localhost port=5432 user=postgres dbname=postgres password=postgres sslmode=disable",
-    "enable_https": true
+    "enable_https": true,
+    "trusted_subnet": "127.0.0.1/24"
 }
