+++ internal/storage/mstorage.go
@@ -208,3 +208,15 @@ func (s *MapStorage) DeleteURL(delURL map[string]uint64) error {
 
 	return nil
 }
+
+func (s *MapStorage) GetStats() (ResJSONStats, error) {
+	urls := len(s.m)
+	users := make(map[uint64]uint64)
+	for _, lURL := range s.m {
+		users[lURL.UserID]++
+	}
+	return ResJSONStats{
+		URLCount:  urls,
+		UserCount: len(users),
+	}, nil
+}
